refactor: import from generation, update cypress tsconfig
