+++ cypress/e2e/builder/invitations/editInvitation.cy.ts
@@ -1,4 +1,4 @@
-import type { PermissionLevel } from '@/openapi/client/types.gen';
+import type { PermissionLevel } from '@/openapi/client';
 
 import {
   ITEM_MEMBERSHIP_PERMISSION_SELECT_CLASS,
