+++ src/modules/builder/components/item/sharing/ItemMembershipSelect.tsx
@@ -70,7 +70,7 @@ const ItemMembershipSelect = ({
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [value]);
 
-  const values = (['read', 'write', 'admin'] as const[]).filter(
+  const values = ['read' as const, 'write' as const, 'admin' as const].filter(
     (p) =>
       value ? allowDowngrade || PermissionLevelCompare.gte(p, value) : true,
   );
