+++ src/modules/builder/contexts/OutletContext.tsx
@@ -1,6 +1,8 @@
 import { createContext, useContext } from 'react';
 
-import { PackedItem, PermissionLevel } from '@graasp/sdk';
+import type { PackedItem } from '@graasp/sdk';
+
+import type { PermissionLevel } from '@/openapi/client';
 
 type OutletContextType = {
   item: PackedItem;
