+++ src/query/mutations/invitation.ts
@@ -1,6 +1,8 @@
-import { Invitation, PermissionLevel, UUID } from '@graasp/sdk';
+import { Invitation, UUID } from '@graasp/sdk';
 
 import { useMutation, useQueryClient } from '@tanstack/react-query';
+
+import type { PermissionLevel } from '@/openapi/client';
 
 import * as Api from '../api/invitation.js';
 import { itemKeys } from '../keys.js';
